Revert "feat: add advanced search to ccc export settings fields in netsuite (#867)"
This reverts commit ce6b051fba73f1d3b23edea5652ea97ecce2c8d1.
