+++ src/app/core/models/intacct/intacct-configuration/export-settings.model.ts
@@ -4,7 +4,6 @@ import { DefaultDestinationAttribute, DestinationAttribute } from "../../db/dest
 import { IntacctDestinationAttribute } from "../db/destination-attribute.model";
 import { SelectFormOption } from "../../common/select-form-option.model";
 import { brandingConfig, brandingContent, brandingFeatureConfig } from "src/app/branding/branding-config";
-import { ExportSettingOptionSearch } from "../../common/export-settings.model";
 
 export type ExportSettingFormOption = {
     label: string,
@@ -56,7 +55,9 @@ export type ExportSettingPost = {
     general_mappings: ExportSettingGeneralMapping
   }
 
-export interface IntacctExportSettingOptionSearch extends ExportSettingOptionSearch {
+export type ExportSettingOptionSearch = {
+    searchTerm: string,
+    destinationAttributes: any[],
     destinationOptionKey: IntacctExportSettingDestinationOptionKey
 }
