+++ src/app/integrations/netsuite/netsuite-shared/netsuite-export-settings/netsuite-export-settings.component.html
@@ -48,8 +48,6 @@
                                 <app-configuration-select-field
                                 [form]="exportSettingForm"
                                 [destinationAttributes]="bankAccounts"
-                                [destinationOptionKey]="NetsuiteExportSettingDestinationOptionKey.BANK_ACCOUNT"
-                                [isOptionSearchInProgress]="isOptionSearchInProgress"
                                 [optionLabel]="'name'"
                                 [iconPath]="'list'"
                                 [isFieldMandatory]="true"
@@ -57,9 +55,7 @@
                                 [label]="brandingContent.corporateCard.defaultDebitCardAccountLabel"
                                 [subLabel]="'Post all your reimbursable expenses to a default employee payables account.'"
                                 [placeholder]="'Select a Employee Payables Account'"
-                                [formControllerName]="'bankAccount'"
-                                [isAdvanceSearchEnabled]="true"
-                                (searchOptionsDropdown)="searchOptionsDropdown($event)">
+                                [formControllerName]="'bankAccount'">
                                 </app-configuration-select-field>
                             </div>
     
@@ -67,8 +63,6 @@
                                 <app-configuration-select-field
                                 [form]="exportSettingForm"
                                 [destinationAttributes]="accountsPayables"
-                                [destinationOptionKey]="NetsuiteExportSettingDestinationOptionKey.ACCOUNTS_PAYABLE"
-                                [isOptionSearchInProgress]="isOptionSearchInProgress"
                                 [optionLabel]="'name'"
                                 [iconPath]="'list'"
                                 [isFieldMandatory]="true"
@@ -76,9 +70,7 @@
                                 [label]="brandingContent.corporateCard.accountsPayableLabel"
                                 [subLabel]="'Post all your reimbursable expenses to a default vendor payables account.'"
                                 [placeholder]="'Select Vendor Payable Account'"
-                                [formControllerName]="'accountsPayable'"
-                                [isAdvanceSearchEnabled]="true"
-                                (searchOptionsDropdown)="searchOptionsDropdown($event)">
+                                [formControllerName]="'accountsPayable'">
                                 </app-configuration-select-field>
                             </div>
                         </div>
@@ -194,8 +186,6 @@
                             <app-configuration-select-field
                             [form]="exportSettingForm"
                             [destinationAttributes]="accountsPayables"
-                            [destinationOptionKey]="NetsuiteExportSettingDestinationOptionKey.ACCOUNTS_PAYABLE"
-                            [isOptionSearchInProgress]="isOptionSearchInProgress"
                             [optionLabel]="'name'"
                             [iconPath]="'list'"
                             [isFieldMandatory]="true"
@@ -203,9 +193,7 @@
                             [label]="brandingContent.corporateCard.accountsPayableLabel"
                             [subLabel]="'Post all your company corporate card transactions to a default vendor payables account.'"
                             [placeholder]="helperService.sentenseCaseConversion('Select Vendor Payable Account')"
-                            [formControllerName]="'accountsPayable'"
-                            [isAdvanceSearchEnabled]="true"
-                            (searchOptionsDropdown)="searchOptionsDropdown($event)">
+                            [formControllerName]="'accountsPayable'">
                             </app-configuration-select-field>
                         </div>
 
@@ -213,8 +201,6 @@
                             <app-configuration-select-field
                             [form]="exportSettingForm"
                             [destinationAttributes]="bankAccounts"
-                            [destinationOptionKey]="NetsuiteExportSettingDestinationOptionKey.BANK_ACCOUNT"
-                            [isOptionSearchInProgress]="isOptionSearchInProgress"
                             [optionLabel]="'name'"
                             [iconPath]="'list'"
                             [isFieldMandatory]="true"
@@ -222,9 +208,7 @@
                             [label]="brandingContent.corporateCard.defaultDebitCardAccountLabel"
                             [subLabel]="'Post all your company corporate card transactions to a default employee payables account.'"
                             [placeholder]="helperService.sentenseCaseConversion('Select a Employee Payables Account')"
-                            [formControllerName]="'bankAccount'"
-                            [isAdvanceSearchEnabled]="true"
-                            (searchOptionsDropdown)="searchOptionsDropdown($event)">
+                            [formControllerName]="'bankAccount'">
                             </app-configuration-select-field>
                         </div>
 
@@ -232,8 +216,6 @@
                             <app-configuration-select-field
                             [form]="exportSettingForm"
                             [destinationAttributes]="cccAccounts"
-                            [destinationOptionKey]="NetsuiteExportSettingDestinationOptionKey.CREDIT_CARD_ACCOUNT"
-                            [isOptionSearchInProgress]="isOptionSearchInProgress"
                             [optionLabel]="'name'"
                             [iconPath]="'list'"
                             [isFieldMandatory]="true"
@@ -241,9 +223,7 @@
                             [label]="brandingContent.corporateCard.defaultCCCAccountLabel"
                             [subLabel]="'Post all your company corporate card transactions to a default credit card account.'"
                             [placeholder]="brandingContent.corporateCard.defaultCCCAccountPlaceholder"
-                            [formControllerName]="'creditCardAccount'"
-                            [isAdvanceSearchEnabled]="true"
-                            (searchOptionsDropdown)="searchOptionsDropdown($event)">
+                            [formControllerName]="'creditCardAccount'">
                             </app-configuration-select-field>
                         </div>
 
@@ -251,8 +231,6 @@
                             <app-configuration-select-field
                             [form]="exportSettingForm"
                             [destinationAttributes]="creditCardVendors"
-                            [destinationOptionKey]="NetsuiteExportSettingDestinationOptionKey.VENDOR"
-                            [isOptionSearchInProgress]="isOptionSearchInProgress"
                             [optionLabel]="'name'"
                             [iconPath]="'list'"
                             [isFieldMandatory]="true"
@@ -260,9 +238,7 @@
                             [label]="brandingContent.corporateCard.defaultCCCVendorLabel"
                             [subLabel]="'Post your company corporate card transactions to a default credit card vendor when a match is not found in NetSuite and auto creation is switched off.'"
                             [placeholder]="'Select default credit card vendor'"
-                            [formControllerName]="'defaultCreditCardVendor'"
-                            [isAdvanceSearchEnabled]="true"
-                            (searchOptionsDropdown)="searchOptionsDropdown($event)">
+                            [formControllerName]="'defaultCreditCardVendor'">
                             </app-configuration-select-field>
                         </div>
